+++ src/services/bind-specs/common.spec.ts
@@ -2,7 +2,8 @@ import { setupTemplate } from '../../mocks/specs.utils'
 
 /** <!> All events are suffixed by the `setupTemplate()` method with unique ids in order to prevent corss^-talk between tests */
 
-describe('Common specs for all data binds', () => {
+// Common specs for all data binds
+describe('Common specs', () => {
 
     beforeEach(done => {
         setupTemplate(`