+++ src/services/bind-specs/deep-nesting.spec.ts
@@ -4,14 +4,15 @@ import { MockWebCmp } from '../../mocks/nano-data-bind.mock'
 
 /** <!> All events are suffixed by the `setupTemplate()` method with unique ids in order to prevent corss^-talk between tests */
 
-describe('Deep nesting of e-data="" data binds', () => {
+// Deep nesting of e-data="" data binds
+describe('Deep nesting', () => {
 
     // <!> Chaining properties in multiple levels from the same web component is possible
     //     However this practice is strongly discouraged.
     //     Developers already have a strong expectation taht a component will receive inputs just from the parent
     //     Receiving inputs (data binds) from other levels than the parent can be hard to read, confusing, and hard to maintan.
     beforeEach(() => setupTemplate(`
-        <mock-web-cmp class="level-1" e-data="mockEvent, { mockValueL1: event.detail }" nano-no-auto-init>
+        <mock-web-cmp class="level-1" e-data="mockEvent, { mockValueL1: event.detail }" no-auto-bind>
             <mock-web-cmp class="level-2" e-data="mockValueL1, { mockValueL2 }">
                 <mock-web-cmp class="level-3" e-data="mockValueL2, { mockValueL3 }">
                 </mock-web-cmp>