+++ src/services/bind-specs/event.spec.ts
@@ -2,7 +2,7 @@ import { setupTemplate } from '../../mocks/specs.utils'
 
 /** <!> All events are suffixed by the `setupTemplate()` method with unique ids in order to prevent corss^-talk between tests */
 
-describe('Origin - Custom event', () => {
+describe('Custom event', () => {
 
     beforeEach(done => {
         setupTemplate(`