+++ src/services/rules-specs/class-rule.spec.ts
@@ -2,7 +2,7 @@ import { setupTemplate, dispatchEvent, id} from '../../mocks/specs.utils'
 
 /** <!> All events are suffixed by the `setupTemplate()` method with unique ids in order to prevent corss^-talk between tests */
 
-describe('Data bind e-class=""', () => {
+describe('Class rule', () => {
     
     beforeEach(done => {
         setupTemplate(`