+++ src/services/rules-specs/multiple-rules.spec.ts
@@ -2,7 +2,7 @@ import { setupTemplate} from '../../mocks/specs.utils'
 
 /** <!> All events are suffixed by the `setupTemplate()` method with unique ids in order to prevent corss^-talk between tests */
 
-describe('Multiple data binds / rules on the same element', () => {
+describe('Multiple rules', () => {
     
     beforeEach(done => { 
         setupTemplate(`